fix: выносит type-guards в отдельный файл
